Move handlers to separate file, change names, export authenticateRequest.
